pyproject.toml
[project]
name = "nonebot-plugin-ai-groupmate"
version = "0.1.0"
description = "miyuki"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "datasets>=3.4.1",
    "expiringdict>=1.2.2",
    "flagembedding>=1.3.4",
    "jieba>=0.42.1",
    "matplotlib>=3.10.1",
    "nb-cli>=1.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.56.2",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-htmlrender>=0.6.5",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-orm>=0.7.7",
    "nonebot-plugin-session>=0.3.2",
    "nonebot-plugin-tortoise-orm>=0.1.4",
    "nonebot-plugin-uninfo>=0.7.0",
    "nonebot2>=2.4.2",
    "numpy>=2",
    "peft>=0.15.0",
    "pillow>=11.1.0",
    "pymilvus[model]>=2.5.6",
    "wordcloud>=1.9.4",
]


[project.urls]
Homepage = "https://github.com/yaowan233/nonebot-plugin-ai-groupmate"
Repository = "https://github.com/yaowan233/nonebot-plugin-ai-groupmate"